Resolve conflicts from merge
