+++ test/test_game.py
@@ -49,12 +49,12 @@ class GameTest(unittest.TestCase):
         result = self.game._board.spaces()
         expected_result = [" " for i in range(9)]
         self.assertEqual(result, expected_result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
-    
+
     def testGamePlayersHaveSeparateSymbolsOnGameStart(self):
         result = self.game._player1._symbol
         expected_result = "X"
         self.assertEqual(result, expected_result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
-        
+
         result = self.game._player2._symbol
         expected_result = "O"
         self.assertEqual(result, expected_result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
@@ -63,6 +63,13 @@ class GameTest(unittest.TestCase):
         result = self.game.current_player()
         expected_result = self.game._player1
         self.assertEqual(result, expected_result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
+    
+    def testGameDisplaysBoardWhenMovePlayed(self):
+        self.game.play_move(self.db)
+
+        result = self.mock_cli_output._last_output
+        expected_result = self.game._output.print_board(self.game._board)
+        self.assertTrue(result in expected_result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
 
     def testCurrentPlayerAfterAFewMovesAreMade(self):
         self.game._board.make_move(5, self.game._player1._symbol)
@@ -79,17 +86,17 @@ class GameTest(unittest.TestCase):
         result = self.game.is_won()
         expected_result = True
         self.assertTrue(result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
-    
+
     def testGameIsWonReturnsFalseWhenGameIsNotWon(self):
         self.draw_game()
 
         result = self.game.is_won()
         expected_result = False
         self.assertFalse(result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
-    
+
     def testGameIsDrawReturnsTrueWhenGameIsADraw(self):
         self.draw_game()
-        
+
         result = self.game.is_draw()
         expected_result = True
         self.assertTrue(result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
@@ -110,7 +117,7 @@ class GameTest(unittest.TestCase):
 
     def testGameReturnsWinningPlayersSymbolXWhenPlayer1HasWon(self):
         self.player1_win()
-        
+
         result = self.game.winner()
         expected_result = self.game._player1._symbol
         self.assertEqual(result, expected_result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
@@ -121,7 +128,7 @@ class GameTest(unittest.TestCase):
         result = self.game.winner()
         expected_result = self.game._player2._symbol
         self.assertEqual(result, expected_result, msg='\nRetrieved:\n{0} \nExpected:\n{1}'.format(result, expected_result))
- 
+
     def testGameDisplaysMessageWhenGameIsWon(self):
         self.player1_win()
         self.game.game_play_loop(self.db)
