+++ tictactoe/game.py
@@ -52,7 +52,6 @@ class Game:
         self._output.print_board(self._board)
 
     def game_play_loop(self, db):
-        self._output.print_board(self._board)
         while not self.is_over():
             self.play_move(db)
